@@ -1,28 +1,35 @@
 {
-  "name": "dormspam",
+  "name": "dormdigest",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://dormspam.github.io/frontend/",
+  "homepage": ".",
   "dependencies": {
     "@material-ui/core": "^4.9.0",
     "@material-ui/icons": "^4.5.1",
-    "axios": "^0.18.1",
+    "@types/react": "^18.0.34",
+    "@types/react-dom": "^18.0.11",
+    "axios": "^1.4.0",
+    "axios-cache-interceptor": "^1.3.0",
     "classnames": "^2.2.6",
+    "http-server": "^14.1.1",
+    "jwk-to-pem": "^2.0.5",
     "material-ui": "^0.20.2",
     "moment": "^2.24.0",
-    "react": "^16.8.0",
+    "react": "^18.2.0",
     "react-data-grid": "^5.0.1",
-    "react-dom": "^16.8.0",
+    "react-dom": "^18.2.0",
     "react-ga": "^2.5.7",
     "react-helmet": "^5.2.0",
-    "react-is": "^16.12.0",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^3.3.0",
+    "react-is": "^18.2.0",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "5.0.1",
     "react-table-6": "^6.11.0",
-    "reactstrap": "^8.4.1"
+    "reactstrap": "^8.4.1",
+    "typescript": "^4.9.5",
+    "universal-cookie": "^6.1.1"
   },
   "scripts": {
+    "TODO": "TODO: fix the package so that it doesn't need this",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
@@ -40,6 +47,10 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "gh-pages": "^2.0.1"
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@types/jwk-to-pem": "^2.0.1",
+    "gh-pages": "^2.2.0",
+    "react-error-overlay": "6.0.9",
+    "sha3": "^2.1.4"
   }
 }
