@@ -1,8 +1,65 @@
-import Request from "./request";
 
+
+const CATEGORIES_COLORS = {
+  'OTHER': "gray",
+  'FOOD': "#dc050c",
+  'CAREER': "#e8601c",
+  'FUNDRAISING': "#f1932d",
+  'APPLICATION': "#f6c141",
+  'PERFORMANCE': "#f7f056",
+  'BOBA': "#cae0ab",
+  'TALKS': "#90c987",
+  'SALE': "#4eb265",
+  'RSVP': "#7bafde"
+}
+
+const CATEGORIES = Object.keys(CATEGORIES_COLORS); //Name of all available categories
 class Categories {
-  static getCategories() {
-    return new Request("/categories");
+  static getCategoriesColorMapping() {
+    //Return mapping of category name to its corresponding color
+    return CATEGORIES_COLORS;
+  }
+  static getCategoriesList() {
+    //Return list of available categories
+    return CATEGORIES;
+  }
+
+  static getCategoriesStringMapping(category_ints) {
+    //Returns list of category names given a list of category integers
+    let category_strings = []
+    for (let category of category_ints) {
+      category_strings.push(CATEGORIES[category]);
+    }
+    return category_strings
+  }
+
+  static getCategoriesIntMapping(category_strings) {
+    //Returns list of category int representations given a list of cateogry strings
+    let category_ints = []
+    for (let category of category_strings) {
+      category_ints.push((CATEGORIES.indexOf(category)));
+    }
+    return category_ints
+  }
+
+  static getCategoriesEmptyFrequency() {
+    //Returns mapping of category name to 0 (to be used as a starting frequency list)
+    const frequency = {};
+    for (const category of CATEGORIES) {
+      frequency[category] = 0;
+    }
+    return frequency;
+  }
+
+  static parseCategoryNameFrequency(categoryNumFrequency) {
+    //Given a object mapping category names => their frequency (doesn't necessarily contain all categories),
+    //return a object mapping category names => their frequency (with all categories)
+    const categoryNameFrequency = this.getCategoriesEmptyFrequency(); //Initialize full category object of zeroes (IMPORTANT)
+
+    for (const [categoryName, categoryFrequency] of Object.entries(categoryNumFrequency)) {
+      categoryNameFrequency[categoryName] = categoryFrequency;
+    };
+    return categoryNameFrequency;
   }
 }
 
