@@ -5,6 +5,7 @@ import Categories from "../../../api/categories";
 import Events from "../../../api/events";
 import HomeFeedEventView from "./HomeFeedEventView";
 import "./HomeFeedView.css";
+import LocalData from "../../../api/localdata";
 
 class HomeFeedView extends Component {
   constructor(props) {
@@ -14,27 +15,26 @@ class HomeFeedView extends Component {
       data: [],
       searchCount: 0,
       searching: false,
-      colors: {}
+      colors: Categories.getCategoriesColorMapping()
     };
 
     this.saveEventData = this.saveEventData.bind(this);
     this.parseCategories = this.parseCategories.bind(this);
-
     const self = this;
 
     Events.getEventsByDate(moment().format("YYYY-MM-DD")).then(response => {
-      self.saveEventData(response.data);
+      self.saveEventData(response);
     });
 
-    Categories.getCategories().then(response => {
-      let tempColors = {};
-      for (let i = 0; i < response.data.length; i++) {
-        tempColors[response.data[i].name] = response.data[i]["color"];
-      }
-      this.setState({
-        colors: tempColors
-      });
-    });
+    // Categories.getCategoriesColorMapping().then(response => {
+    //   let tempColors = {};
+    //   for (let i = 0; i < response.data.length; i++) {
+    //     tempColors[response.data[i].name] = response.data[i]["color"];
+    //   }
+    //   this.setState({
+    //     colors: tempColors
+    //   });
+    // });
   }
 
   componentWillReceiveProps(nextProps) {
@@ -43,21 +43,21 @@ class HomeFeedView extends Component {
     const searchCount = this.state.searchCount + 1;
 
     if (searching) {
+      // filter based on title
       // make axios request here for search
-      Events.getEventsByQuery(nextProps.search).then(response => {
-        if (searchCount < self.state.searchCount) {
-          return;
-        }
+      // Events.getEventsByQuery(nextProps.search).then(response => {
+      //   if (searchCount < self.state.searchCount) {
+      //     return;
+      //   }
 
-        self.saveEventData(response.data);
-      });
+      //   self.saveEventData(response.data);
+      // });
     } else {
       Events.getEventsByDate(nextProps.selectedDay.format("YYYY-MM-DD")).then(response => {
         if (searchCount < self.state.searchCount) {
           return;
         }
-
-        self.saveEventData(response.data);
+        self.saveEventData(response);
       });
     }
 
@@ -66,12 +66,35 @@ class HomeFeedView extends Component {
       searching: searching
     });
   }
+  saveEventData(eventData) {
+    eventData.events.forEach((event, index) => {
+      if (eventData.tags[index]) {
+        const event_tags = eventData.tags[index]; //Name of categories associated with the event
+        event.tags = event_tags;
+        event.user_email = eventData.users[index]; //Email of the user who sent/submitted the event
+        event.description = eventData.descriptions[index]; //Plaintext email description
+        event.description_html = eventData.descriptions_html[index]; //HTML email description
+      }
+    });
 
-  saveEventData(inputData) {
-    console.log(inputData)
+    //Getting a integer representation of the categories the user has selected since event.tags is a list of ints
+    let current_categories = Categories.getCategoriesIntMapping(LocalData.getCategoryFilters());
+
+    //Creating a new list matchingEvents that contains events that have categories matching with the categories the user has selected
+    const matchingEvents = eventData.events.filter((event) => {
+
+      //Creating a list of current_categories(categories that user wants to be shown) that are also in the event tags the event has
+      let matching_categories = current_categories.filter(cat => event.tags.includes(cat));
+
+      //If this list has an element, adding the event to the list of matchingEvents
+      return matching_categories.length > 0;
+    })
+
+    const events = matchingEvents;
+
     let times = [];
 
-    let data = inputData.sort((a, b) => {
+    let data = events.sort((a, b) => {
       let aTime = moment(a.start_time).valueOf();
       let bTime = moment(b.start_time).valueOf();
 
@@ -84,7 +107,8 @@ class HomeFeedView extends Component {
       return 0;
     });
 
-    for (var i=0; i < data.length; i++) {
+
+    for (var i = 0; i < data.length; i++) {
       if (times.length === 0) {
         times.push([data[i]]);
       } else {
@@ -95,34 +119,38 @@ class HomeFeedView extends Component {
         }
       }
     }
-
-    for (var j=0; j < data.length; j++) {
-      data[j].categories = this.parseCategories(data[j].categories);
+    for (var j = 0; j < data.length; j++) {
+      data[j].tags = this.parseCategories(data[j].tags);
     }
 
-    let filteredTimes = [];
-
-    for (var k=0; k < times.length; k++) {
-      for (var l=0; l < times[k][0].categories.length; l++) {
-        if (this.props.categories.indexOf(times[k][0].categories[l]) >= 0) {
-          filteredTimes.push(times[k]);
-          break;
-        }
-      }
-    }
+    // let filteredTimes = [];
+    //
+    // for (var k=0; k < times.length; k++) {
+    //   for (var l=0; l < times[k][0].tags.length; l++) {
+    //     if (this.props.categories.indexOf(times[k][0].tags[l]) >= 0) {
+    //       filteredTimes.push(times[k]);
+    //       break;
+    //     }
+    //   }
+    // }
 
     this.setState({
-      data: filteredTimes
+      data: times
     });
   }
 
+  // parseCategories(categories) {
+  //   categories = categories.substring(1, categories.length - 1);
+  //   let listed = categories.split(",");
+  //   for (let i=0; i < listed.length; i++) {
+  //     listed[i] = listed[i].replace(/"/g, "");
+  //   }
+  //   return listed;
+  // }
+
   parseCategories(categories) {
-    categories = categories.substring(1, categories.length - 1);
-    let listed = categories.split(",");
-    for (let i=0; i < listed.length; i++) {
-      listed[i] = listed[i].replace(/"/g, "");
-    }
-    return listed;
+    const tags = Categories.getCategoriesList();
+    return categories.map(category => tags[category]);
   }
 
   render() {
@@ -136,12 +164,12 @@ class HomeFeedView extends Component {
           </div>
         </div>
       );
+      let timeString = moment(`${this.state.data[i][0].start_date}T${this.state.data[i][0].start_time}`)
+      timeString = timeString.format("h:mm a");
 
-      let timeString = moment(this.state.data[i][0].start_time).format("h:mm a");
-
-      if (this.state.searching) {
-        timeString = moment(this.state.data[i][0].start_time).format("MMMM Do YYYY, h:mm a");
-      }
+      // if (this.state.searching) {
+      //   timeString = moment(this.state.data[i][0].start_time).format("MMMM Do YYYY, h:mm a");
+      // }
 
       elements.push(
         <div className="onetime" key={"times" + i + "2"}>{timeString}</div>
@@ -150,11 +178,12 @@ class HomeFeedView extends Component {
         let selected = false;
 
         if (this.props.selectedEvent !== null) {
-          selected = this.props.selectedEvent.uid === this.state.data[i][j].uid;
+          selected = this.props.selectedEvent.id === this.state.data[i][j].id;
         }
 
+
         elements.push(
-          <div className="timeevents" key={this.state.data[i][j].uid}>
+          <div className="timeevents" key={this.state.data[i][j].id}>
             <div className="sidespace" />
             <HomeFeedEventView
               event={this.state.data[i][j]}
@@ -166,12 +195,11 @@ class HomeFeedView extends Component {
       }
     }
 
-    console.log(elements)
-    if (elements.length === 0){
+    if (elements.length === 0) {
       elements = <div className="empty-events">
-                  <img className="empty-events-icon" src="img/empty.png"/>
-                  <h1 className="empty-events-label">No events found!</h1>
-                </div>
+        <img className="empty-events-icon" src="img/empty.png" />
+        <h1 className="empty-events-label">No events found!</h1>
+      </div>
     }
 
     return (
